+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## Unrelease `0.14.0`
 
 - **Added**
+
   - `#198`: As scp user, I want to opt for a prompt if `username` is missing. (`option.prompt`)
   - `#201`: As scp user, I want to opt for a prompt if `password` is missing.
   - `#204`: As scp user, I want to opt for a prompt if `host` is missing.
   - `#202`: As scp user, I want to opt for a prompt if `localPath` is missing.
+  - `#203`: As scp user, I want to opt for a prompt if `remotePath` is missing.
+
 - **Test**
   - `#208`: As test, I want to refactor `username` tests to use `mockResolvedValueOnce`